Extract `Item` component
