+++ frontend/src/metabase/containers/ItemPicker/ItemPicker.jsx
@@ -20,13 +20,9 @@ import EntityListLoader, {
 import { entityObjectLoader } from "metabase/entities/containers/EntityObjectLoader";
 
 import Collections from "metabase/entities/collections";
-import {
-  ItemContent,
-  ExpandItemIcon,
-  ItemPickerHeader,
-  ItemPickerList,
-  ItemRoot,
-} from "./ItemPicker.styled";
+
+import Item from "./Item";
+import { ItemPickerHeader, ItemPickerList } from "./ItemPicker.styled";
 
 const getCollectionIconColor = () => color("text-light");
 
@@ -281,49 +277,3 @@ export default _.compose(
     getCollectionIcon: (props.entity || Collections).objectSelectors.getIcon,
   })),
 )(ItemPicker);
-
-const Item = ({
-  item,
-  name,
-  icon,
-  color,
-  selected,
-  canSelect,
-  hasChildren,
-  onChange,
-  onChangeParentId,
-}) => {
-  const iconProps = _.isObject(icon) ? icon : { name: icon };
-  return (
-    <ItemRoot
-      mt={1}
-      p={1}
-      onClick={
-        canSelect
-          ? () => onChange(item)
-          : hasChildren
-          ? () => onChangeParentId(item.id)
-          : null
-      }
-      canSelect={canSelect}
-      isSelected={selected}
-      hasChildren={hasChildren}
-      data-testid="item-picker-item"
-    >
-      <ItemContent>
-        <Icon size={22} {...iconProps} color={selected ? "white" : color} />
-        <h4 className="mx1">{name}</h4>
-        {hasChildren && (
-          <ExpandItemIcon
-            name="chevronright"
-            canSelect={canSelect}
-            onClick={e => {
-              e.stopPropagation();
-              onChangeParentId(item.id);
-            }}
-          />
-        )}
-      </ItemContent>
-    </ItemRoot>
-  );
-};
